Fix PHP deprecations
