+++ tests/ClosureSerializer/SuperClosureSerializerTest.php
@@ -9,7 +9,7 @@ class SuperClosureSerializerTest extends TestCase
 {
     public function setUp(): void
     {
-        if (! class_exists(\SuperClosure\SerializerInterface::class)) {
+        if (! interface_exists(\SuperClosure\SerializerInterface::class)) {
             $this->markTestSkipped('Missing jeremeamia/superclosure to run this test');
         }
     }
